refactor(netsync): Replication.hpp becomes five headers, and consumers say what they use
1,904 lines declaring both halves of the protocol, the interfaces a game plugs
into it, and the config that tunes it — so every consumer that wanted any one of
those pulled in all of them.

  ReplicationConfig.hpp     relevancy/replication config and the loader
  ReplicationProviders.hpp  the interfaces a game implements, and diagnostics
  InstanceRecord.hpp        a received instance, and the expander that builds it
  ReplicationServer.hpp     the sending half
  ReplicationClient.hpp     the receiving half

Replication.hpp is gone rather than left as an umbrella: all 21 consumers now
include the headers they actually use, which is most of the point. Half of them
turn out to want two or three, not five — DistanceRelevancy.hpp needs the
provider interface and the config and nothing else, and the server TUs split
three ways.

DistanceRelevancy.cpp was reaching ReplicationServer through the umbrella
without naming it, and now includes it.

One edit that is not a move: the client's `@brief` had drifted above
InstanceRecord, so InstanceRecord carried two @brief lines and ReplicationClient
none. It is back on the class it describes.

Build clean, ctest 14/14.
