+++ modules/NetSync/CMakeLists.txt
@@ -10,8 +10,12 @@ target_sources(Assisi-NetSync
     "include/Assisi/NetSync/NetClock.hpp"
     "include/Assisi/NetSync/NetComponents.hpp"
     "include/Assisi/NetSync/NetProtocol.hpp"
+    "include/Assisi/NetSync/InstanceRecord.hpp"
     "include/Assisi/NetSync/NetSession.hpp"
-    "include/Assisi/NetSync/Replication.hpp"
+    "include/Assisi/NetSync/ReplicationClient.hpp"
+    "include/Assisi/NetSync/ReplicationConfig.hpp"
+    "include/Assisi/NetSync/ReplicationProviders.hpp"
+    "include/Assisi/NetSync/ReplicationServer.hpp"
   PRIVATE
     "src/BodyState.cpp"
     "src/DistanceRelevancy.cpp"
