+++ modules/NetSync/include/Assisi/NetSync/NetSession.hpp
@@ -28,7 +28,9 @@
 #include <Assisi/Net/NetTransport.hpp>
 #include <Assisi/NetSync/InputCommand.hpp>
 #include <Assisi/NetSync/NetClock.hpp>
-#include <Assisi/NetSync/Replication.hpp>
+#include <Assisi/NetSync/ReplicationClient.hpp>
+#include <Assisi/NetSync/ReplicationConfig.hpp>
+#include <Assisi/NetSync/ReplicationServer.hpp>
 
 #include <cstdint>
 #include <memory>
