+++ modules/NetSync/tests/TestContentSet.cpp
@@ -28,7 +28,8 @@
 #include <Assisi/ECS/Transform.hpp>
 #include <Assisi/Net/NetTransport.hpp>
 #include <Assisi/NetSync/NetComponents.hpp>
-#include <Assisi/NetSync/Replication.hpp>
+#include <Assisi/NetSync/ReplicationClient.hpp>
+#include <Assisi/NetSync/ReplicationServer.hpp>
 
 using namespace Assisi;
 using namespace Assisi::NetSync;
