+++ modules/NetSync/tests/TestControl.cpp
@@ -23,7 +23,8 @@
 #include <Assisi/ECS/Transform.hpp>
 #include <Assisi/Net/NetTransport.hpp>
 #include <Assisi/NetSync/NetComponents.hpp>
-#include <Assisi/NetSync/Replication.hpp>
+#include <Assisi/NetSync/ReplicationClient.hpp>
+#include <Assisi/NetSync/ReplicationServer.hpp>
 
 #include <cstdint>
 #include <vector>
