+++ structured_workflow.txt
@@ -1,0 +1,176 @@
+ISSUE WORKFLOW
+==============
+
+Two roles. The orchestrator is the main session; it never does the engineering
+work itself. The worker is a subagent that does all of it.
+
+
+CONTEXT YOU WILL NOT HAVE
+-------------------------
+
+Assisi is a C++ game engine. These issues are all remediation on one long-lived
+branch:
+
+  - Work happens on 'blueprints', not main and not dev. PRs merge into
+    'blueprints'. Do not target anything else.
+
+  - The worklist of record is docs/blueprint-review-round7-findings.md,
+    section 9. Read §9 before starting anything. It supersedes §7, it reorders
+    the work into tiers (§9.2), it lists which findings depend on each other
+    (§9.3), and §9.5 drops the findings that turned out to be wrong. Issues cite
+    finding IDs — B9, S26, T12 — and those are looked up in that document.
+    Fixing something §9.5 already dropped is how this branch wasted effort once.
+
+  - Linear: team "Engine", project "Blueprint Remediation". Estimates are in
+    points.
+
+  - The Linear issues are synced to GitHub issues. Every Linear issue carries a
+    link to its GitHub twin in its attachments — that is where the GitHub issue
+    number comes from. Read it off the issue; never compute it from the ENG
+    number. Because of that sync, anything done on the GitHub side shows up in
+    Linear on its own. Do not hand-write Linear comments or hand-set issue
+    states to report something GitHub already knows.
+
+  - Tests are doctest. Build with `make gd`; run everything with `ctest` from
+    out/build/gcc-debug, or run one module's binary directly, e.g.
+    out/build/gcc-debug/modules/App/tests/Assisi-App-Tests -tc="some case*".
+
+  - A test that documents a bug which is not fixed yet is marked
+    `doctest::should_fail()`. That keeps the suite green while the bug stands,
+    and it inverts the moment the bug is fixed — doctest reports "should have
+    failed but didn't" and the run goes red until the decorator is removed. Use
+    it when an issue asks for coverage of a defect whose fix belongs to a
+    different issue, and say so in a comment naming the finding. Never assert
+    something a correct fix could not deliver; check first that the assertion
+    is satisfiable.
+
+  - The orchestrator's notification tool is PushNotification. It is a deferred
+    tool: load it with ToolSearch("select:PushNotification") before calling it.
+
+
+ROLES
+-----
+
+ORCHESTRATOR
+  - Picks the issue and hands it to a worker.
+  - Spawns exactly one worker per issue, on Opus, with fresh context.
+  - Talks to that worker with SendMessage for the whole life of the issue.
+    Never spawns a second worker for the same issue.
+  - Retires the worker once the issue is closed. The next issue gets a new one.
+  - Sends the push notification and holds the approval gate.
+
+WORKER
+  - Does every step below.
+  - May NOT spawn subagents of its own.
+  - May NOT push, open a PR, or merge anything before the programmer approves.
+    Nothing reaches the remote before that.
+
+
+ORDER OF ISSUES
+---------------
+
+Roughly by issue number, but it is not a rule. Dependencies win: if an issue is
+blocked, or another has to land first, do whatever order actually works.
+
+
+PART 1 — SETUP (worker)
+-----------------------
+
+ 1. Assign the issue to Francisco Vivas (DaFrancc) in Linear. The user is
+    already in the workspace.
+ 2. Set the estimate in Linear.
+ 3. Read the issue description in full.
+ 4. Get the branch name from the issue (Linear supplies it, e.g. feature/eng-98),
+    and the GitHub issue number from the issue's attachments.
+ 5. Fetch first, and make sure local 'blueprints' is level with the remote. The
+    branch is cut from it, and a stale 'blueprints' cuts from the wrong tip.
+    Then create the branch as a linked branch on that GitHub issue:
+
+      gh issue develop <github-issue-number> --base blueprints \
+          --name <branch-name-from-linear> --checkout
+
+    One command creates it off 'blueprints', pushes it, links it to the issue,
+    and checks it out. Do not create the branch with git and push it by hand:
+    that produces a branch GitHub does not associate with the issue, so the
+    association never reaches Linear.
+ 6. Confirm the link took, with `gh issue develop --list <github-issue-number>`.
+    The link is the announcement — GitHub carries it to Linear on its own. Do
+    not post a Linear comment or move the issue state to say the branch exists.
+
+    This is the one remote action allowed before approval; it publishes an empty
+    branch, nothing else. `gh issue develop` cannot adopt a branch that is
+    already on the remote, so if an unlinked one is sitting there, delete it
+    (`git push origin --delete <branch-name>`, which leaves local commits alone)
+    and then run the command above.
+
+
+PART 2 — THE WORK (worker)
+--------------------------
+
+ 7. Make a plan. The programmer does NOT need to approve it; it exists so the
+    worker goes in with one.
+ 8. If the plan needs a decision only the programmer can make, stop and ask
+    before continuing.
+ 9. Write the test cases first, where a test would prove the bug exists or is
+    genuinely needed for correctness. Do not write vacuous tests — a test that
+    passes whether or not the behaviour works is worse than no test. Confirm
+    each new test fails when the behaviour it names is broken.
+10. Do the work.
+11. Verify it. Building is not verifying: run the tests.
+12. Commit. The final commit title must contain "Fixes ENG-#", which closes the
+    Linear issue on merge.
+
+
+PART 3 — THE GATE
+-----------------
+
+13. WORKER: stop here and report to the orchestrator. Say what was done, what
+    the commit is, and any setback or surprise worth knowing. Do not push.
+14. ORCHESTRATOR: send a push notification with a short, plain summary and any
+    setbacks. Keep it brief — no essays. Then wait.
+15. PROGRAMMER: may ask questions. Approval is what releases the rest.
+
+
+PART 4 — LANDING IT (worker, resumed by SendMessage after approval)
+-------------------------------------------------------------------
+
+16. Push the commits.
+17. Open a PR with gh, merging into 'blueprints'.
+18. Title the PR "Fixes ENG-#". Same reason as the commit title — belt and
+    braces on closing the Linear issue.
+19. The PR has no description.
+20. Set the PR assignee in gh to DaFrancc. If that is not possible, skip it.
+21. Merge the PR.
+22. Delete the feature branch, locally and on the remote.
+23. Pull the merge into local 'blueprints': check it out, fast-forward it from
+    origin, and confirm the work is actually in it — not just merged on GitHub.
+    Leave the repo sitting on an up-to-date 'blueprints', which is what the next
+    issue branches from.
+24. ORCHESTRATOR: retire the worker and move to the next issue without asking.
+
+
+WHEN TO INTERRUPT THE PROGRAMMER
+--------------------------------
+
+Only three times:
+  - A problem in the plan.
+  - A problem in the implementation.
+  - The approval gate at step 13.
+
+Nothing else. Assume the programmer is away from the terminal, and send the
+notification even if they appear to be present.
+
+
+CONVENTIONS
+-----------
+
+  - No commit trailers of any kind. No Co-Authored-By, no session URL. The
+    programmer is the sole author.
+  - Explicit-width integer types: int32_t, uint32_t, and so on. Never bare int
+    or unsigned.
+  - Never add a `throw`. Catch what dependencies throw at our boundary and
+    return std::expected.
+  - Build with the make targets (make gd, make gcc-debug), not raw cmake.
+  - Never use sed, not even to page a file. Use the file tools.
+  - Churn is affordable. Do not scope a refactor down to spare call sites.
+  - Do not edit plan or design documents without an explicit go-ahead.
